sysfs: Add open/release to bin_attribute
In order to support runtime pm on binary attributes that access devices,
add open and release callbacks to struct bin_attribute. This will give the
option to keep the device awake for as along as the file is opened. A
better option may have been to use vma ops to perform this action at the
time the driver is mmaped, but due to commit a6849fa1f7d7 ("sysfs: Fail bin
file mmap if vma close is implemented."), it is not possible to use
vma->ops->close().

Signed-off-by: David E. Box <david.e.box@linux.intel.com>