sched/fair: Shorten balancing interval
When a CPU transitions from idle to busy, there are interactions between
the idle and periodic load balancer triggers that result in the value
rq->next_balance being continuously pushed into the future. The result
is that a CPU that just became busy may not pull high-throughput tasks
onto itself for up to 1 second.

Mitigate this problem by not considering a CPU busy state when computing
the next balance time.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>