verify the support of mba4
MBA4 is a new added feature in kernel for X86 platform, we're
intending to add this test case to validate it. Briefly MBA4.0
is a extension of legacy MBA, with the same interface, but more
flexible function. This patch is mainly a sanity testing, more
function testing will be in following patch.

Case1: Sanity test for resctrl FS:
 mount resctrl FS with mba4, and check hardware capability.

Case2: Negative test for resctrl FS:
 mount resctrl FS without mba4, and check hardware capability.

Reported-by: kernel test robot <lkp@intel.com>
Reviewed-by: Yi Sun <yi.sun@intel.com>
Signed-off-by: You Zhou <you.zhou@intel.com>
Signed-off-by: Tony Luck <tony.luck@intel.com>