verify the function of mba4
The major difference between MBA3 and MBA4 is that the bandwidth of MBA4 is
actually a threshold other than a limitation, so need to add test cases to
cover this new feature.

Case3: under competition
A process with a higher percentage bandwidth means that it will take higher
priority to complete when it's under competition. Both MBA3 and MBA4 would
have the same behavior in this case.

Case4: under non-competition
if a task running without memory competition, it would run with unlimited
bandwidth, so we compare the diff between unlimit_bandwidth_time and
limit_bandwidth_time, The case is successful, if the diff is small. Or the
case is failed.

Reported-by: kernel test robot <lkp@intel.com>
Reviewed-by: Yi Sun <yi.sun@intel.com>
Signed-off-by: You Zhou <you.zhou@intel.com>
Signed-off-by: Tony Luck <tony.luck@intel.com>