idxd: Work around SWQ retry issue
SWQ may be shared to multiple uses so that the queue may be full
quickly. This will block other requests. To avoid this, we design
to send retry back to guest when migrating the mdev. But something
wrong in the retry flow so that guest cannot correctly send retry
out.

To unblock the test, make this work around to queue the desc as
DWQ. When final solution found, may revert this patch.

Signed-off-by: Yi Sun <yi.y.sun@linux.intel.com>