Added non-x86 forward declare for svos_cmds_init
This was caught in a PowerPC build of the Intel Next kernel. The non-x86 path
has a definition for svos_cmds_init but no prototype for it. The x86 path has
a forward declaration for it, so we mimick that for the non-x86 path too.

This should fix an error in Intel Next branch internal-eywa/master.
Commit: 60753d0ff190c347574742615e1f9f6b4ec420fe

Reported-by: kernel test robot <lkp@intel.com>