dmaengine: idxd: set priv bit to 0 when wq is not kernel type
Set priv bit to 0 for read-only WQ setup when the WQ type is not kernel.
Current code does not touch the priv bit when the WQ is not kernel type.
This non-action causes problem wq reconfiguration where the WQ is
configured as a kernel type, gets disabled, and then reconfigured to be a
user WQ. The priv bit that was set by the kernel type config remains.

Signed-off-by: Dave Jiang <dave.jiang@intel.com>