platform/x86/intel/sdsi: Add attestation and measurement support
Intel On Demand hardware adds attestation and firmware measurement
retrieval services through use of SPDM 1.0 [1] protocols. The attestation
flow is used to attest the authenticity of On Demand hardware through
verification flow based on a stored certificate chain. The measurement flow
allows retrieval and verification of both the state certificate and the
meter certificate.  These allow software to verify the authenticity of the
On Demand hardware as well as the integrity of the reported silicon
configuration.

Add an SPDM transport for sending SPDM messages through the On Demand
mailbox. Use generic netlink ABI with commands to find devices, authorize
(attest) and to get measurements. The root key to use for verification is
to be added separately by the user to the kernel key management system.

[1] https://www.dmtf.org/sites/default/files/standards/documents/DSP0274_1.0.1.pdf

Signed-off-by: David E. Box <david.e.box@linux.intel.com>
Link: https://www.dmtf.org/sites/default/files/standards/documents/DSP0274_1.0.1.pdf