drm/i915/lnl: Introduce MDCLK
In LunarLake we now separate MDCLK from CDLCK, which used to be before
always 2 times CDCLK.  Now we might afford lower CDCLK, while having
higher memory clock, so improving bandwidth and power consumption same
time.  This patch is part of preparations required, to enable that.

v2: Fix checkpatch and remove comment providing and alternative approach
that will not going to be implemented (Lucas)

Signed-off-by: Stanislav Lisovskiy <stanislav.lisovskiy@intel.com>
Reviewed-by: Lucas De Marchi <lucas.demarchi@intel.com>