drm/i915/lnl: Introduce MDCLK_CDCLK ratio to DBuf
When we change MDCLK/CDCLK the BSpec now instructs us to
write a ratio between MDCLK/CDCLK to MBUS CTL and DBUF CTL
registers during that change.
Previsouly DBuf state and CDCLK were not anyhow coupled together,
now at compute stage when we know which CDCLK/MDCLK we are going
to use, we need to update the DBuf state with that ratio, being
properly encoded, so that it gets written to those registers,
once DBuf state is being update.
The criteria for updating DBuf state is also a CDCLK/MDCLK ratio
change now.

v2: Fixed mistake with uninitialized ratio variable
v3 (Lucas):
  - Remove d15_ prefix
  - Fix and move display version check
  - Keep convention of newest platform first in
    get_mbus_mdclk_cdclk_ratio()

Signed-off-by: Stanislav Lisovskiy <stanislav.lisovskiy@intel.com>
Acked-by: Lucas De Marchi <lucas.demarchi@intel.com>
Reviewed-by: Mika Kahola <mika.kahola@intel.com>