DEBUG: drivers: thermal: hfi: Create /proc/classid
Create the file /proc/classid to read the classification value that
the hardware provides for the current thread in the current CPU.

In effect, it reads the MSR_IA32_HW_FEEDBACK_CHAR MSR and returns
the classid if valid.

Applications can use this file to perform work and then read the
classification that hardware provided.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>