pci: doe: introduce a new synchronous message exchange interface
This new interface pci_doe_message_exchange_sync() is for the case,
when user has already prepared DOE header for request and expect
DOE header for the response (e.g. for length checking). So this is
a specific use case of the DOE mailbox, especially in TDX-IO/SPDM
case, that trusted firmware generates message with DOE header
directly. So far SPDM layer is not implemented, so such function
is put into DOE mailbox code for now, but can be move to SPDM part
once SPDM layer is ready.

Signed-off-by: Wu Hao <hao.wu@intel.com>
Signed-off-by: Ming Li <ming4.li@intel.com>