KVM: TDX: add a simple model for private iommu to support private dev
Ideally the private iommu implementation should be in iommu driver, but
it is WIP. So this patch uses a simple model for KVM to track:

1. How many iommu devices is being used in KVM across TDs.
2. How many iommu devices is being used by a TD.
3. Which iommu device is the private device connected to.

These info are useful when performing cache invalidation. Since the
iommu Second Stage Paging is shared with EPT, so when the EPT changes,
KVM should do iotlb invalidation for all affected iommus. Also a iommu
device should avoid concurrent cache invalidation operations to ensure
the correctness of its Invalidation Queue.

Signed-off-by: Xu Yilun <yilun.xu@intel.com>