tdx-filter: Authorize GED driver to probe
The goal of this patch is to allow a TDX guest to probe the evged
driver so that a PCI device can be hotplugged through ACPI.

First it adds the ability to allow platform devices in addition to
the existing support for ACPI and PCI devices.

Then it adds the GED device ACPI0013 as an authorized device for
both ACPI and platform buses.

It modifies the platform core code by setting the device as authorized
in case it was part of the authorized list maintained by the TDX filter.

And last thing is to update the ACPI IO remapping code to make it rely
on the hardened version of ioremap(). The reason for this is to allow
IO resources declared by authorized ACPI devices to be accessed.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>