x86/uintr: Introduce vector registration and uvec_fd syscall
Each receiving task has its own interrupt vector space of 64 vectors.
For each vector registered by a task create a uvec_fd. Only tasks that
have previously registered a user interrupt handler can register a
vector.

The sender for the user interrupt could be another userspace
application, kernel or an external source (like a device). Any sender
that wants to generate a user interrupt needs access to receiver's
vector number and UPID.  uvec_fd abstracts that information and allows
a sender with access to uvec_fd to connect and generate a user
interrupt. Upon interrupt delivery, the interrupt handler would be
invoked with the associated vector number pushed onto the stack.

Using an FD abstraction automatically provides a secure mechanism to
connect with a receiver. It also makes the tracking and management of
the interrupt vector resource easier for userspace.

uvec_fd can be useful in some of the usages where eventfd is used for
userspace event notifications. Though uvec_fd is nowhere close to a
drop-in replacement, the semantics are meant to be somewhat similar to
an eventfd or the write end of a pipe.

Access to uvec_fd can be achieved in the following ways:
- direct access if the task is part of the same thread group (process)
- inherited by a child process.
- explicitly shared using any of the FD sharing mechanisms.

If the sender is another userspace task, it can use the uvec_fd to send
user IPIs to the receiver. This works in conjunction with the SENDUIPI
instruction. The details related to this are covered later.

The exact APIs for the sender being a kernel or another external source
are still being worked upon. The general idea is that the receiver would
pass the uvec_fd to the kernel by extending some existing API (like
io_uring).

The vector associated with uvec_fd can be unregistered by closing all
references to the uvec_fd.

Signed-off-by: Sohil Mehta <sohil.mehta@intel.com>