x86/uintr: Add support for interrupting blocking system calls
User interrupts are delivered to applications immediately if they are
running in userspace. Without this support, if a receiver task has
blocked in the kernel due to blocking calls like read(), sleep(), etc;
the interrupt will only get delivered when the application gets
scheduled again.

To avoid that, add support for interrupting tasks that are blocking when
a user interrupt is triggered. This has similar semantics to how signals
work today.

There are 2 options implemented here:
1) The cpu that was last running the task handles the kernel interrupt
and wakes up the receiver task. There is no hardware assist here to
identify the receiver. So, there is a need to scanning all tasks that were
blocked.

2) Set a reserved bit in the UPID which would cause a #GP on the
sender task trying to generate the interrupt. In the #GP handler, decode
and emulate the instruction to wake up the receiver.

<Both of the above options are experimental and the implemented code
susceptible to races.>

Signed-off-by: Sohil Mehta <sohil.mehta@intel.com>