x86/tdx: Virtualize CPUID leaf 0x2
CPUID leaf 0x2 provides cache and TLB information. In TDX guest access
to the leaf causes #VE.

Current implementation returns all zero, but it confuses some users:
GLIBC crashes inside qsort_r(3). It is likely a GLIBC bug, but it is
also user-visible regression comparing to non-TDX environment.

Kernel has to generate a sensible response to the #VE.

The leaf is obsolete. There are leafs that provides the same
information in a structured form. See leaf 0x4 on cache info and
leaf 0x18 on TLB info.

Generate a response that indicates that CPUID leaf 0x4 and 0x18 have to
be used instead.

Signed-off-by: Kirill A. Shutemov <kirill.shutemov@linux.intel.com>