KVM: TDX: add the seamcall TDH_MIG_STREAM_CREATE
The TDH_MIG_STREAM_CREATE seamcall creates a migration stream and its
migsc control structure in the TDX module. Migration stream is a TDX
concept and it is analogous to a "channel" in TDX for later TD states
export and import.

The seamcall can be invoked at anytime after the tdcs pages have been
added for the TD, and before a migration session starts.

Signed-off-by: Wei Wang <wei.w.wang@intel.com>