KVM: TDX: add seamcall TDH_EXPORT_BLOCKW
The TDH_EXPORT_BLOCKW seamcall blocks a list of TD private pages for
writing (i.e. write-protecting). The list of private pages are provided
by the gpa_list_info, which is in the format below (defined in TDX
module ABI spec):
Bits 2:0   - Reserved (must be 0)
Bits 11:3  - Index of the first entry in the GPA list page to process
Bits 51:12 - Host physical address of the GPA list page, which holds up
             to 512 64-bit guest-physical addresses. The GPA list page
             must be a shared page.
Bits 54:52 - Reserved (must be 0)
Bits 63:55 - Index of the last entry in the GPA list page to process

If a page failed to be blocked for writing, the error is marked in the
related entry of the GPA list page, but the seamcall doesn't abort, and
the number of pages that failed will be includes in the returned status
(i.e. in rax).

TDH_EXPORT_BLOCKW is interruptible. If a pending interrupt is detected
during operation, it returns TDX_INTERRUPTED_RESUMABLE with rcx holding
the next entry in the GPA list page to continue. The caller can re-invoke
the seamcall immediately after handling the interrupt.

Signed-off-by: Wei Wang <wei.w.wang@intel.com>