KVM: TDX_MIG: allow userspace to map the buffers
On the migration source, the userspace migration thread needs to read the
migration data exposed from the TDX module in the KVM allocated buffers
to send to the destination. On the migration destination, the userspace
migration thread needs to writes the received migration data into the
KVM allocated buffers to import into the TDX module.

Allow the userspace to map the buffers for direct accesses. To be more
precise, those buffers include:
- MBMD buffer: stores the migration bundle metadata
- GPA buffer: allocated for the GPA list and stores the GPA entries
- MAC buffer: allocated for the MAC list and stores MACs
- Memory buffers: allocated for the memory buffer list and stores the
  private memory data

Signed-off-by: Wei Wang <wei.w.wang@intel.com>