KVM: TDX_MIG: add seamcall TDH_IMPORT_STATE_IMMUTABLE
The TDH_IMPORT_STATE_IMMUTABLE seamcall starts a new migration session
on the destination and imports the TD's immutable states. The imported
data includes the MBMD data and immutable states data generated by
TDH_EXPORT_STATE_IMMUTABLE on the source.

The mbmd_info field provides the buffer that stores MBMD, and its format
is defined as below in the TDX module ABI spec:
Bits 51:0  - Host physical address of the buffer, including HKID bits
Bits 63:52 - Size of the buffer

The page_list_info field provides the list of 4KB pages that stores the
TD-scope immutable state data.

The mig_stream_info field provides the info about the migration stream
(should have been created through TDH_MIG_STREAM_CREATE ) from the TDX
module to load the state. Its format:
Bits 15:0  - migration stream index (must be 0)
Bits 62:16 - reserved (must be 0)
Bits 63    - Indicates if this is resumption of a previously interrupted
             operation (being 1) or a new invocation (being 0).

The seamcall is interruptible, and the caller is expected to invoke it
in a loop when the returning status is TDX_INTERRUPTED_RESUMABLE, and
the above resumption flag (bit 63) of mig_stream_info needs to be set.

Signed-off-by: Wei Wang <wei.w.wang@intel.com>