acpica: add acpi_os_map_memory_opregion
In confidential computing scenarios operation regions
need to be shared with the host in order for the
host to pass data via such regions to the guest.
Since by default all guest mappings are private, define
a new acpi_os_map_memory_opregion interface that maps
operation regions as shared in case we are executing
in a confidential guest. This avoids mapping rest of the
ACPI structures as shared and only shares the operation
regions.

Signed-off-by: Elena Reshetova <elena.reshetova@intel.com>