KVM: TDX: add TDG_VP_VMCALL_SERVICE
TDG_VP_VMCALL_SERVICE is a sub function of TDG.VP.VMCALL used by a service
TD to communicate with KVM. The tdvmcall_service struct is defined to as
the format for the communication data (unified the definition of Table
3-42 and 3-43 defined in GHCI spec). The 128-bit guid identifies the
service types, and the data field points to the service specific data.

When making the TDG_VP_VMCALL, the service TD allocates a 4KB aligned
command buffer, and a 4KB aligned response buffer shared with KVM. The
Guest-physical addresses of the 2 buffers are given via R12 and R13
registers. The command buffer is filled by the service TD with commands
for KVM to handle, and the response buffer is filled by the KVM to respond
to the service TD. The various command and response types are documented
in Chapter 3.13 in the TDX Guest Host Communication Interface (GHCI) spec.

When receiving the TDG.VP.VMCALL, KVM allocates 2 host buffers of the
same size as the command buffer and response buffer, and copies the
commands into the host side buffer. When the command handling is done, the
response data inthe KVM allocated response buffer are copied to the
service TD shared response buffer. This avoids the inconvinence of direct
accessing to userspace memory.

The current implementation doesn't support interrupt to service TD,
meaning that the service TD is expected to poll on responses. If the
response isn't ready, the service TD can be halted by KVM and be waken
up later when response is ready to read.

Signed-off-by: Wei Wang <wei.w.wang@intel.com>