drivers: thermal: intel: hfi: Add Meteorlake to class post-processing
Meteorlake behaves like Alderlake and Raptorlake. IPC classification is
more reliable if only one SMT sibling is busy at a given time. It is not
clear if classes 2 and 3 can be reliable identified in Meteorlake.

Only use the classification that hardware provides if the SMT sibling of
a CPU is idle.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>