KVM: TDX: add seamcall TDH_SYS_RD
The TDH_SYS_RD seamcall is added in TDX module v1.5 to allow KVM to read a
TDX module global-scope metadata field. It can be used by KVM to enumerate
the TDX Module's capabilities and configuration ( e.g. the max number of
service TDs that the current version of TDX module supports for a user
TD to be bound to).

Availability of TDH_SYS_RD is indicated by the SYS_RD boolean flag added
to TDSYSINFO_STRUCT (TDX ABI spec, Table 4.18), so the seamcall is made
only when the SYS_RD flag is reported as true.

Signed-off-by: Wei Wang <wei.w.wang@intel.com>