KVM: TDX: read the max number of servtds supported per user TD
During TDX module setup, read from the TDX module the max number of
service TDs supported for a user TD to be bound to. This is done by
passing the field ID of the max_servtd field as defined in the TDX ABI
spec with the TDH.SYS.RD seamcall.

Signed-off-by: Wei Wang <wei.w.wang@intel.com>