KVM: TDX_MIG: allow userspace to set parameters to kvm_device
The buf_list_pages is the only parameter for userspace to set to
kvm_device currently. buf_list is a list of pages to export or import TD
states from the TDX module. Userspace tells KVM the list size (i.e. number
of pages) that it will use, for example, userspace could do a batch of 12
TD private pages to export. The buf_list is also used to export/import
TD non-memory states (e.g. immutable states, vCPU state), so the size
should be larger or equal to the size required by the non-memory states.
The max number of pages supported by the TDX architecture is 512
currently, so userspace is expected to request a number smaller than 512.

Signed-off-by: Wei Wang <wei.w.wang@intel.com>