drm/i915/bmg: Reduce register timeout to prevent
The display register timeout (RM timeout) releases stuck
transactions on RM bus. The timeout is programmable, but
defaults to 65ms, which is so high that SoC machine check
(cat error) crash is hit before the display timeout can
release a stuck transaction. It is a security problem for
display register accesses, particularly reads, to cause a
system crash, so the timeout needs to be reduced to avoid
crashes.

Hitting the display timeout indicates something is wrong
in hardware, software, or firmware.  The timeout should
not be relied on for proper display behavior. Display
may be internally broken with corrupted or blank
screen if a timeout is hit, but the system should not
crash.

WA: 14012195489

--v2
Removed defaulting to 2ms as already RM_TIMEOUT default
value configured to 1.5 ms. Attached Bspec link.
[Suraj]

https://gfxspecs.intel.com/Predator/
ContentEdit/(S(xgtl4jczmbw3w4lmvh03o2vq))/
Content/Bun/GEN_HAS_14012195489

--v3 to --v9
Patch Rebase to resolve Sporadic issues

--v10
Limited Workaround implementation to Display Version 14

Signed-off-by: Mitul Golani <mitulkumar.ajitkumar.golani@intel.com>
Reviewed-by: Suraj Kandpal <suraj.kandpal@intel.com>