drm/i915/xe2lpd: read pin assigment from IOM for display version >= 20
Starting from display version 20, we need to read the pin assignment
from the IOM TCSS_DDI_STATUS register instead of reading it from the
FIA.

We use the pin assignemnt to decide the maximum lane count.  So, to
support this change, add a new lnl_tc_port_get_max_lane_count()
function that reads from the TCSS_DDI_STATUS register and decides the
maximum lane count based on that.

BSpec: 69594
Reviewed-by: Mika Kahola <mika.kahola@intel.com>
Reviewed-by: Lucas De Marchi <lucas.demarchi@intel.com>
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>