KVM: TDX: add initialized check to is_td/vcpu_created
TD creation or vcpu creation is complete only when td or vcpu has been
initialized. For the destination TD during live migration, the td
finishes the creation after the TD-scope state is imported (where
td_initialized is set to true then), and the vcpu finishes the creation
after the vcpu states are imported (where the vcpu_initialized is set
to true).

Signed-off-by: Wei Wang <wei.w.wang@intel.com>