drm/xe/sim: Skip minimal loading of GuC for reading hwconfig
Loading GuC fw twice is causing GuC failures in simulator. For the first
time, the GuC fw is loaded with minimal ads configuration, to read the
hwconfig, then later again with full ads.
The minimal loading of GuC to read the hwconfig is skipped if GuC_RESET
is disabled to allow the module load to pass till the issue is resolved.

HSD to track in Cobalt:
https://hsdes.intel.com/appstore/article/#/16019817552

Signed-off-by: Balasubramani Vivekanandan <balasubramani.vivekanandan@intel.com>
Reviewed-by: Lucas De Marchi <lucas.demarchi@intel.com>