KVM: TDX: always allow userspace to access PV MSRs
Userspce may not know which PV MSRs is emulated for TDX. Just allow
userspace to access all the PV MSRs, as they might need to be migrated
from the source to the destination. The MSRs that are not enabled for
the TD are just set with an unused value. This doesn't matter.

Signed-off-by: Wei Wang <wei.w.wang@intel.com>