drm/xe/sim: Define a bitmap of features to be disabled in presilicon
There are certain features which needs to be disabled when executing on
presilicon environment because the feature may not be available on the
hardware model. The features to be disabled varies across the presilicon
environments like simulation, different emulation models and also
depends on the platform.
A bitmap is defined which holds the features to be disabled and is
initialized during the device probe based on the platform and the
execution environment.
The convenient macros are created which can be used throughout the
driver to check if a particular feature is enabled before using it.

Signed-off-by: Balasubramani Vivekanandan <balasubramani.vivekanandan@intel.com>