KVM: TDX: clear td_initialized and vcpu_initialized flags on destroy
After the tdvpx and tdvpr pages are reclaimed, the TD vcpu becomes not
initialized, so clear the vcpu_initialized flag. After the tdcs and tdr
pages are reclaimed, TD is not initialized, so clear the TD's initialized
flag.

Some functions run based on the check of the "initialized" flag and could
be called after the pages have been reclaimed. Having the flags cleared
avoid wrong checks in this case.

Signed-off-by: Wei Wang <wei.w.wang@intel.com>