KVM: Introduce a new flag for memory slot allowing page conversion in non-UPM case
In current UPM implementation, it allows private memfd backed memory
slots to do page conversion. While for the device assignment, the MMIO
access can encounter page conflict as well. Whether to allow the page
conversion for such MMIO access by default is still uncertain, because:
- UPM defaults to shared and guest MMIO access with shared address
  directly. Such conversion handling is unncessary in current design.
- But TDXIO would require such capability to convert page attribute of
  pass-thru device MMIO region.

To make it more flexible, introduce a flag KVM_MEM_NONUPM_SAFE to
indicate the target memory slot is MMIO region of pass-thru device. It is
allowed to do page conversion for such kind of slot if set. It is worth
noting that KVM_MEM_NONUPM_SAFE and KVM_MEM_PRIVATE is mutually
exclusive.

KVM_MEM_NONUPM_SAFE means the slot can be private, but cannot be private
faultin because of no gmem support.

Signed-off-by: Chenyi Qiang <chenyi.qiang@intel.com>
Signed-off-by: Xu Yilun <yilun.xu@intel.com>