drm/xe/xe2: Add GT topology readout
Xe2 platforms have three DSS fuse registers for both geometry and
compute.

FIXME: The registers in the topology file should be moved to the
dedicated register header after the tree is rebased.

FIXME2:  The XEHPC_ prefix on the third compute register may need to be
XE2 if RLT gets ZBB'd.

Signed-off-by: Matt Roper <matthew.d.roper@intel.com>