tdxio: add DEVIFMT support
Implement the DEVIF Metadata Table (DEVIFMT) management based on the
common matadata table interfaces for TDXIO.

A set of common matadata table interfaces is introduced to serve both
DEVIFMT and MMIOMT. They are constructed with similar data structures.

DEVIFMT is a tree map structure of TDI's FUNCION_ID to DEVIFCS. It is
to enforce that each TDI is represented by a single DEVIFCS page. It
should be created before a corresponding DEVIFCS is created.

Signed-off-by: Xu Yilun <yilun.xu@intel.com>