platform/x86/intel/pmt: Add API to read fixed telemetry
Some PMT telemetry implementations provide access to telemetry that is
guaranteed to be "fixed". This means that the metrics are at fixed offsets
that will not change (though may be extended) in future products. This for
hardcoded descriptions in the kernel that do not need to be updated.

Fixed telemetry is denoted by the "telem_type" field which is currently not
used. Add this field and add an API to allows registering for a fixed
telemetyr endpoint.

Signed-off-by: David E. Box <david.e.box@linux.intel.com>