merge.py: Fix __contains__ bug
Dictionaries and lists have different semantics for __contains__.

If I do `k in list`, that iterates through the list to look
for an element.

If I do `k in dict`, that examines the key set. These are not the
same!

Instead, we just switch to a standard try-except structure. It's
more Pythonic anyway.