fixup! drm/xe/sim: Add parameter to force presi mode
Call presi init hook earlier in the device probe.
This is done to allow the manipulation of the platform feature flags for
simulation.
So instead of changing the feature flags in the pci xe_device_desc
structure, it can enabled/disabled inside the presilicon init
hook. So the presilicon changes remain isolated from the actual driver
code.

Signed-off-by: Balasubramani Vivekanandan <balasubramani.vivekanandan@intel.com>