tools/power/turbostat: Add initial support for new platform probing mechanism
Different pm features that turbostat supports diverge on different CPU
Models.

Currently, CPU Model is referenced in different places to handle this
divergency. This introduces the code complexity, and also introduces the
difficulty to add completed support for each new model.

In order to cleanup the various references of CPU Model ID, a new
feature based platform probing mechanism is introduced.

First of all, two concepts are introduced first,
1. Feature ID : a unique value to represnet a specific feature. Only
   feature that diverges among platforms need a Feature ID.
2. Feature value : defines the behavior for a Feature ID. It can be
   bool: describes if a Feature ID is supported or not
   unique value: describes a certain behavior for a Feature ID
   bitmap: describes a set of subfeatures supported for a Feature ID

Based on these two concepts,
1. introduce enum feature_id to represent unique Feature IDs.
2. introduce global array pm_features[] to cache the Feature Value for
   each Feature ID.
3. initialize pm_features[] array based on CPU model at launch time.
4. introduce fid_mask to ensure that every Feature ID is assigned a
   Feature Value explicitly.

Depends on the feature, Feature value for each Feature ID can be
declared and used in three different ways as below,
1. features that have two states: enabled/disabled.
   use enable_feature(FID_FOO) to disable the feature.
   use disable_feature(FID_FOO, 1) to enabled the feature.
   use
      if (get_feature(FID_FOO))
   to check if the feature is supported or not.

2. features that have multiple behaviors.
   use disable_feature(FID_FOO) to disable the feature.
   use set_feature(FID_FOO, X) to described the behavior.
   use
      if (get_feature(FID_FOO) == X)
   to check if the feature behaves in an expected way or not.

3. features that is composed of a series of subfeatures (cstate, etc).
   use disable_feature(FID_FOO) to disable the feature.
   use set_feature(FID_FOO, X | Y | Z) to described the list of
      subfeatures supported.
   use
      if (get_feature(FID_FOO) & X)
   to check if a subfeature is supported or not.

The initial support is introduced in this patch, code will be cleaned up
in later patches, and the CPU model lists will be removed one by one
when all its reference is cleaned up.

In the end, CPU model will be referenced only once during turbostat
launch time. And the platform support for each Feature IDs will be set a
Feature value explicitly, in a central place, when the CPU model is
referenced.

Signed-off-by: Zhang Rui <rui.zhang@intel.com>