KVM: x86: Disable migration for kvm_x86_vendor_init()
Disable migration for kvm_x86_vendor_init() so that all hardware
information can be fetched on the same cpu, as a preparation to move
VMXON out of KVM.

For example, setup_vmcs_config() called in hardware_setup() is supposed
to only read MSRs from the same cpu, but it is called when preemption is
enabled.

To completely solve this problem, we need to convert all boot_cpu_has()
to cpu_has(), and disable migration during the entire progress.  For now
only do the latter so that it's safe to use per-cpu variable and
smp_processor_id() during hardware_setup().

Link: https://lore.kernel.org/lkml/ZAdxNgv0M6P63odE@google.com/T/#m153008637806732461be461934947a7ae3464c0f
Signed-off-by: Kai Huang <kai.huang@intel.com>