KVM: x86: Advertise AVX-VNNI-INT16 CPUID to user space
AVX-VNNI-INT16 is a new set of instructions in the latest Intel platform
Clearwater Forest, for media decoding, image post-processing.

The bit definition:
CPUID.(EAX=7,ECX=1):EDX[bit 10]

AVX-VNNI-INT16 is on a KVM-only subleaf. Plus an x86_FEATURE definition
for this feature bit to direct it to the KVM entry.

Advertise AVX-VNNI-INT16 to KVM userspace. This is safe because there
are no new VMX controls or additional host enabling required for guests
to use this feature.

Signed-off-by: Tao Su <tao1.su@linux.intel.com>