tools/power/turbostat: Introduce FID_RAPL_QUIRK_TDP
Introduce FID_RAPL_QUIRK_TDP and use it to abstract if a hardcoded
30Watts TDP should be used when it cannot be retrieved from
MSR_PKG_POWER_UNIT.

Remove unused CPU model check in get_tdp_intel() at the same time.

Note that, when FID_RAPL_QUIRK_TDP disabled, a 135Watts TDP will be used
when it cannot be retrieved from MSR_PKG_POWER_UNIT. This does not seem
to be right neither. But this is only a cornor case, and how to fix it
is a separate work, thus it is not covered here.

Signed-off-by: Zhang Rui <rui.zhang@intel.com>